requirements.txt
# Core dependencies
flask==3.1.1
flask-cors==5.0.0
Werkzeug==3.1.3

# WSGI Server
gunicorn==23.0.0

# Development dependencies
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
alembic==1.17.1

# Security
pyotp==2.9.0
cryptography==43.0.1
flask-limiter==3.10.0
flask-talisman==1.1.0
python-jose[cryptography]==3.3.0

# Caching & Sessions
redis==5.2.1
flask-caching==2.3.0
flask-session==0.8.0

# Monitoring & Observability
prometheus-flask-exporter==0.23.1
sentry-sdk[flask]==2.21.0
python-json-logger==3.2.1

# API Documentation
flasgger==0.9.7.1
apispec==6.7.1

# Validation
pydantic==2.10.5
marshmallow==3.24.1

# Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-asyncio==0.24.0
pytest-mock==3.14.0
factory-boy==3.3.1
faker==33.1.0

# Code Quality
black==24.10.0
flake8==7.1.1
bandit==1.8.0
pre-commit==4.0.1
